update deps and add routing groups
