+++ package.json
@@ -17,7 +17,7 @@
   "devDependencies": {
     "@iconify-json/heroicons": "^1.1.4",
     "@julr/unocss-preset-forms": "^0.0.2",
-    "@mpaupulaire4/rollup-plugin-pages": "^1.0.4",
+    "@mpaupulaire4/rollup-plugin-pages": "^2.0.0",
     "@sveltejs/vite-plugin-svelte": "^1.0.1",
     "@tsconfig/svelte": "^3.0.0",
     "@typescript-eslint/eslint-plugin": "^5.33.1",
