+++ src/pages/(auth)/signup/+page.svelte
@@ -3,25 +3,20 @@
   import { createForm } from 'felte';
 
   import Input from 'components/forms/Input.svelte';
-  import { resetPassConfirm, resetPassRequest } from 'lib/users/auth';
+  import { register } from 'lib/users/auth';
   import { title } from '../title';
 
-  interface FormData {
+  type FormData = {
     email: string;
     password: string;
-    passwordConfirm: string;
-  }
-
-  const token = new URLSearchParams(location.search).get('token');
+    confirmPassword: string;
+  };
 
   const { route } = useRouter();
 
   const { form } = createForm<FormData>({
     onSubmit: data => {
-      if (token) {
-        return resetPassConfirm(token, data.password, data.passwordConfirm);
-      }
-      return resetPassRequest(data.email);
+      return register(data.email, data.password, data.confirmPassword);
     },
     onError: e => {
       // show error message
@@ -33,34 +28,29 @@
       route('/');
     },
   });
-  title.set('Reset Passwoard');
+  title.set('Sign up for an account');
 </script>
 
 <form class="space-y-6" use:form>
-  {#if token}
-    <Input
-      type="text"
-      label="New Passwoord"
-      name="password"
-      autocomplete="new-password"
-      required
-    />
-    <Input
-      type="text"
-      label="Confirm Password"
-      name="passwordConfirm"
-      autocomplete="new-password"
-      required
-    />
-  {:else}
-    <Input type="text" label="Email Address" name="email" autocomplete="email" required />
-  {/if}
-
+  <Input type="text" label="Email Address" name="email" autocomplete="email" required />
+  <Input
+    type="password"
+    label="Password"
+    name="password"
+    autocomplete="new-password"
+    required
+  />
+  <Input
+    type="password"
+    label="Confirm Password"
+    name="confirmPassword"
+    autocomplete="new-password"
+    required
+  />
   <div class="flex items-center justify-between">
-    <div />
     <div class="text-sm">
-      <a href="/auth/login" class="font-medium text-indigo-600 hover:text-indigo-500">
-        Need to log in?
+      <a href="/login" class="font-medium text-indigo-600 hover:text-indigo-500">
+        Already have an account?
       </a>
     </div>
   </div>
