+++ src/pages/(protected)/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import svelteLogo from '../assets/svelte.svg';
+  import svelteLogo from 'assets/svelte.svg';
 </script>
 
 <main>
